+++ packages/agent/test/routes/modification/dissociate-delete-related.test.ts
@@ -96,7 +96,12 @@ describe('DissociateDeleteRelatedRoute', () => {
         await count.handleDissociateDeleteRelatedRoute(context);
 
         expect(dataSource.getCollection('bookPersons').update).toHaveBeenCalledWith(
-          { email: 'john.doe@domain.com', requestId: expect.any(String), timezone: 'Europe/Paris' },
+          {
+            email: 'john.doe@domain.com',
+            requestId: expect.any(String),
+            request: { ip: expect.any(String) },
+            timezone: 'Europe/Paris',
+          },
           new PaginatedFilter({
             conditionTree: factories.conditionTreeBranch.build({
               aggregator: 'And',
@@ -164,6 +169,7 @@ describe('DissociateDeleteRelatedRoute', () => {
             {
               email: 'john.doe@domain.com',
               requestId: expect.any(String),
+              request: { ip: expect.any(String) },
               timezone: 'Europe/Paris',
             },
             new PaginatedFilter({
@@ -230,6 +236,7 @@ describe('DissociateDeleteRelatedRoute', () => {
               {
                 email: 'john.doe@domain.com',
                 requestId: expect.any(String),
+                request: { ip: expect.any(String) },
                 timezone: 'Europe/Paris',
               },
               new PaginatedFilter({
@@ -352,7 +359,12 @@ describe('DissociateDeleteRelatedRoute', () => {
         await count.handleDissociateDeleteRelatedRoute(context);
 
         expect(dataSource.getCollection('librariesBooks').delete).toHaveBeenCalledWith(
-          { email: 'john.doe@domain.com', requestId: expect.any(String), timezone: 'Europe/Paris' },
+          {
+            email: 'john.doe@domain.com',
+            requestId: expect.any(String),
+            request: { ip: expect.any(String) },
+            timezone: 'Europe/Paris',
+          },
           new Filter({
             conditionTree: factories.conditionTreeBranch.build({
               aggregator: 'And',
@@ -419,6 +431,7 @@ describe('DissociateDeleteRelatedRoute', () => {
             {
               email: 'john.doe@domain.com',
               requestId: expect.any(String),
+              request: { ip: expect.any(String) },
               timezone: 'Europe/Paris',
             },
             new Filter({
@@ -484,6 +497,7 @@ describe('DissociateDeleteRelatedRoute', () => {
               {
                 email: 'john.doe@domain.com',
                 requestId: expect.any(String),
+                request: { ip: expect.any(String) },
                 timezone: 'Europe/Paris',
               },
               new Filter({
@@ -590,7 +604,12 @@ describe('DissociateDeleteRelatedRoute', () => {
         await count.handleDissociateDeleteRelatedRoute(context);
 
         expect(dataSource.getCollection('bookPersons').delete).toHaveBeenCalledWith(
-          { email: 'john.doe@domain.com', requestId: expect.any(String), timezone: 'Europe/Paris' },
+          {
+            email: 'john.doe@domain.com',
+            requestId: expect.any(String),
+            request: { ip: expect.any(String) },
+            timezone: 'Europe/Paris',
+          },
           new Filter({
             conditionTree: factories.conditionTreeBranch.build({
               aggregator: 'And',
@@ -655,6 +674,7 @@ describe('DissociateDeleteRelatedRoute', () => {
             {
               email: 'john.doe@domain.com',
               requestId: expect.any(String),
+              request: { ip: expect.any(String) },
               timezone: 'Europe/Paris',
             },
             new Filter({
@@ -713,6 +733,7 @@ describe('DissociateDeleteRelatedRoute', () => {
               {
                 email: 'john.doe@domain.com',
                 requestId: expect.any(String),
+                request: { ip: expect.any(String) },
                 timezone: 'Europe/Paris',
               },
               new Filter({
@@ -831,7 +852,12 @@ describe('DissociateDeleteRelatedRoute', () => {
         await count.handleDissociateDeleteRelatedRoute(context);
 
         expect(dataSource.getCollection('librariesBooks').delete).toHaveBeenCalledWith(
-          { email: 'john.doe@domain.com', requestId: expect.any(String), timezone: 'Europe/Paris' },
+          {
+            email: 'john.doe@domain.com',
+            requestId: expect.any(String),
+            request: { ip: expect.any(String) },
+            timezone: 'Europe/Paris',
+          },
           new Filter({
             conditionTree: factories.conditionTreeBranch.build({
               aggregator: 'And',
@@ -859,7 +885,12 @@ describe('DissociateDeleteRelatedRoute', () => {
         );
 
         expect(dataSource.getCollection('libraries').delete).toHaveBeenCalledWith(
-          { email: 'john.doe@domain.com', requestId: expect.any(String), timezone: 'Europe/Paris' },
+          {
+            email: 'john.doe@domain.com',
+            requestId: expect.any(String),
+            request: { ip: expect.any(String) },
+            timezone: 'Europe/Paris',
+          },
           new Filter({
             conditionTree: factories.conditionTreeBranch.build({
               aggregator: 'And',
@@ -937,6 +968,7 @@ describe('DissociateDeleteRelatedRoute', () => {
             {
               email: 'john.doe@domain.com',
               requestId: expect.any(String),
+              request: { ip: expect.any(String) },
               timezone: 'Europe/Paris',
             },
             new Filter({
@@ -969,6 +1001,7 @@ describe('DissociateDeleteRelatedRoute', () => {
             {
               email: 'john.doe@domain.com',
               requestId: expect.any(String),
+              request: { ip: expect.any(String) },
               timezone: 'Europe/Paris',
             },
             new Filter({
@@ -1041,6 +1074,7 @@ describe('DissociateDeleteRelatedRoute', () => {
               {
                 email: 'john.doe@domain.com',
                 requestId: expect.any(String),
+                request: { ip: expect.any(String) },
                 timezone: 'Europe/Paris',
               },
               new Filter({
@@ -1065,6 +1099,7 @@ describe('DissociateDeleteRelatedRoute', () => {
               {
                 email: 'john.doe@domain.com',
                 requestId: expect.any(String),
+                request: { ip: expect.any(String) },
                 timezone: 'Europe/Paris',
               },
               new Filter({