+++ packages/plugin-export-advanced/test/index.test.ts
@@ -19,6 +19,8 @@ describe('plugin-export-advanced', () => {
             fields: {
               id: factories.columnSchema.numericPrimaryKey().build(),
               title: factories.columnSchema.build({ columnType: 'String' }),
+              isPublished: factories.columnSchema.build({ columnType: 'Boolean' }),
+              publishedAt: factories.columnSchema.build({ columnType: 'Date' }),
               author: factories.manyToOneSchema.build({
                 foreignCollection: 'authors',
                 foreignKey: 'id',
@@ -29,11 +31,15 @@ describe('plugin-export-advanced', () => {
             {
               id: 1,
               title: "The Hitchhiker's Guide to the Galaxy, vol 2",
+              isPublished: true,
+              publishedAt: new Date(),
               author: { id: 1, fullname: 'Douglas Adams' },
             },
             {
               id: 2,
               title: 'The Restaurant at the End of the Universe',
+              isPublished: false,
+              publishedAt: new Date(),
               author: null,
             },
           ],
@@ -92,11 +98,11 @@ describe('plugin-export-advanced', () => {
           watchChanges: false,
         },
         {
-          enumValues: ['id', 'title', 'author:id', 'author:fullname'],
+          enumValues: ['id', 'title', 'isPublished', 'publishedAt', 'author:id', 'author:fullname'],
           id: 'Fields',
           label: 'Fields',
           type: 'EnumList',
-          value: ['id', 'title', 'author:id', 'author:fullname'],
+          value: ['id', 'title', 'isPublished', 'publishedAt', 'author:id', 'author:fullname'],
           watchChanges: true,
         },
       ]);
@@ -107,13 +113,16 @@ describe('plugin-export-advanced', () => {
       ['.json', 'application/json'],
       ['.xlsx', 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'],
     ])('running the action should export to %s', async (format, mimeType) => {
-      const result = await dataSource
-        .getCollection('books')
-        .execute(factories.caller.build(), 'Export books (advanced)', {
+      const result = await dataSource.getCollection('books').execute(
+        factories.caller.build(),
+        'Export books (advanced)',
+        {
           Format: format,
           Filename: 'file',
-          Fields: ['id', 'title', 'author:id', 'author:fullname'],
-        });
+          Fields: ['id', 'title', 'isPublished', 'publishedAt', 'author:id', 'author:fullname'],
+        },
+        factories.filter.build(),
+      );
 
       expect(result.type).toStrictEqual('File');
       expect((result as FileResult).mimeType).toStrictEqual(mimeType);
@@ -156,7 +165,7 @@ describe('plugin-export-advanced', () => {
     test('running the action should export to csv', async () => {
       const result = await dataSource
         .getCollection('books')
-        .execute(factories.caller.build(), 'Export', {});
+        .execute(factories.caller.build(), 'Export', {}, factories.filter.build());
 
       expect(result.type).toStrictEqual('File');
       expect((result as FileResult).mimeType).toStrictEqual('text/csv');